infra/modules/cloud_functions/src/requirements.txt
functions-framework==3.*
requests
google-cloud
google-cloud-pubsub
os
